Make the approximate stratified sampler more flexible
It's now possible to not sample from every subgroup in each batch. That
is, if we have 60 classes and 6 subgroups, then before, the minimum
batch size was 360, but with the new setting, it can be just 60: each
class is present in each batch but only one (uniformly-randomly sampled)
subgroup per class is present.